Cleanup old profile types
